+++ CHANGELOG.md
@@ -35,6 +35,8 @@ with dynamic choices in this way.
 * Prevent un-publishing the `@apostrophecms/global` doc and more generally all singletons.
 * When opening a context menu while another is already opened, prevent from focusing the button of the first one instead of the newly opened menu.
 * Updates `isEqual` method of `area` field type to avoid comparing an area having temporary properties with one having none.
+* In a relationship field, when asking for sub relationships using `withRelationships` an dot notion. 
+If this is done in combination with a projection, this projection is updated to add the id storage fields of the needed relationships for the whole `withRelationships` path. 
 
 ## 4.7.2 and 4.8.1 (2024-10-09)